doctory/requirements.txt
Django==3.2
djangorestframework>=3.12.1
django-cors-headers>=3.7
django-extensions>=3.1.0
psycopg2-binary>=2.8
coverage>=5.5
whitenoise==5.2.0